Worked mainly on exception (server, client side).
Introduced Response to Exception (client) and Exception to Response (server) mappers.
First one convert client response to exceptions if needed; those will be catched in the
Resource handler.
Second convert exceptions into response; it has been chosen to always send back a JSON
output with 200 HTTP code, with the error code and the error message being part of the reply.